refactor(core): split delta module helpers
